Re organize into helpers/constants file
